Add guide for concatenating selectors
By using string concatenation, the real selectors no longer exist in the
codebase's CSS/Sass, making them hard to search for. When someone does
search for a selector which is concatenated during build time, their
search might come up empty which gives the impression that it is unused
code and it could be mistakingly removed.